Autoconfiscate
Add a GNU autotools build system, so that the library can be installed in
the standard way, and the project be used as a sub-project of Beetle.

Use libtool-rs to generate a .la file for the library. With the current
version of libtool-rs, this only works for x86_64 Linux.

Cargo.toml is templated by autoconf, to avoid repetition.